Fix: change online status
There was a little mistake in the last version. The online status was not indicated correctly in the plugin info. With this change we fix this problem, right now, when we change the plugin status (online/offline) this status is reflected correctly.

Additionally, we added the offline status as a language variable.